Improve compatibility with old behavior
Up to version 0.8.0, there were errors on processing valid current Unicode code points so, for old code bases, it was introduced the constant XKBCOMMON_UNICODE_COMPAT that preserves the old behavior when it is 1 or adheres to strict valid current Unicode code points when it is 0.

The current default value for XKBCOMMON_UNICODE_COMPAT is 1 but it is recommended to set it to 0 on build systems so that new code is future proof. The default value may change in future versions.